moved music specific tasks into music namespace, more code review feedback changes
