+++ web-app/test/sidekiq/music/artist_description_job_test.rb
@@ -13,7 +13,7 @@ class Music::ArtistDescriptionJobTest < ActiveSupport::TestCase
     mock_result.stubs(:success?).returns(true)
     mock_task.stubs(:call).returns(mock_result)
 
-    Services::Ai::Tasks::ArtistDescriptionTask.expects(:new).with(parent: @artist).returns(mock_task)
+    Services::Ai::Tasks::Music::ArtistDescriptionTask.expects(:new).with(parent: @artist).returns(mock_task)
 
     @job.perform(@artist.id)
   end
@@ -24,7 +24,7 @@ class Music::ArtistDescriptionJobTest < ActiveSupport::TestCase
     mock_result.stubs(:success?).returns(true)
     mock_task.stubs(:call).returns(mock_result)
 
-    Services::Ai::Tasks::ArtistDescriptionTask.stubs(:new).returns(mock_task)
+    Services::Ai::Tasks::Music::ArtistDescriptionTask.stubs(:new).returns(mock_task)
     Rails.logger.expects(:info)
 
     @job.perform(@artist.id)
@@ -37,7 +37,7 @@ class Music::ArtistDescriptionJobTest < ActiveSupport::TestCase
     mock_result.stubs(:error).returns("AI service unavailable")
     mock_task.stubs(:call).returns(mock_result)
 
-    Services::Ai::Tasks::ArtistDescriptionTask.stubs(:new).returns(mock_task)
+    Services::Ai::Tasks::Music::ArtistDescriptionTask.stubs(:new).returns(mock_task)
     Rails.logger.expects(:error)
 
     @job.perform(@artist.id)
@@ -47,7 +47,7 @@ class Music::ArtistDescriptionJobTest < ActiveSupport::TestCase
     Music::Artist.expects(:find).with(@artist.id).returns(@artist)
 
     # Mock the task to avoid actual execution
-    Services::Ai::Tasks::ArtistDescriptionTask.stubs(:new).returns(mock.tap { |m| m.stubs(:call).returns(mock.tap { |r| r.stubs(:success?).returns(true) }) })
+    Services::Ai::Tasks::Music::ArtistDescriptionTask.stubs(:new).returns(mock.tap { |m| m.stubs(:call).returns(mock.tap { |r| r.stubs(:success?).returns(true) }) })
 
     @job.perform(@artist.id)
   end
